+++ AppboyKit/headers/AppboyKitLibrary/AppboyKit.h
@@ -2,6 +2,7 @@
 #import "ABKUser.h"
 #import "ABKAppboyEndpointDelegate.h"
 #import "ABKFacebookUser.h"
+#import "ABKTwitterUser.h"
 
 // Feedback
 #import "ABKFeedbackViewController.h"