+++ Example/StopwatchDevelop WatchKit Extension/NotificationController.m
@@ -1,0 +1,35 @@
+
+#import "NotificationController.h"
+
+
+@interface NotificationController()
+
+@end
+
+
+@implementation NotificationController
+
+- (instancetype)init {
+    self = [super init];
+    if (self){
+    }
+    return self;
+}
+
+- (void)willActivate {
+    [super willActivate];
+}
+
+- (void)didDeactivate {
+    [super didDeactivate];
+}
+
+- (void)didReceiveRemoteNotification:(NSDictionary *)remoteNotification withCompletion:(void (^)(WKUserNotificationInterfaceType))completionHandler {
+  NSLog(@"remote notification going to be presented on watch: %@", remoteNotification);
+    completionHandler(WKUserNotificationInterfaceTypeDefault);
+}
+
+@end
+
+
+